# swagger.yaml
{"openapi":"3.0.1","info":{"title":"Microservice Postal Item","description":"Сервис почтовых отправлений"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Postal Item","description":"Сервис почтовых отправлений"}],"paths":{"/postalitem/newstatus":{"put":{"tags":["Postal Item"],"description":"Изменение статуса почтового отправления","operationId":"postalItemChangeStatus","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusDto"}}},"required":true},"responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostalDelivery"}}}},"401":{"description":"Unauthorized"}}}},"/postoffice/create":{"post":{"tags":["Postal Item"],"description":"Создание нового почтового отделения","operationId":"createPostOffice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostOfficeDto"}}},"required":true},"responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostOffice"}}}},"401":{"description":"Unauthorized"}}}},"/postalitem/create":{"post":{"tags":["Postal Item"],"description":"Создание и регистрация нового почтового отправления","operationId":"createPostalItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostalDeliveryDto"}}},"required":true},"responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostalDelivery"}}}},"401":{"description":"Unauthorized"}}}},"/statushistory/{id}":{"get":{"tags":["Postal Item"],"description":"Получение истории и статусов почтового отправления по id","operationId":"getStatusHistoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"}}}},"/postoffice/{id}":{"get":{"tags":["Postal Item"],"description":"Получение почтового отделения по id","operationId":"getPostOfficeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostOffice"}}}},"401":{"description":"Unauthorized"}}}},"/postoffice/all":{"get":{"tags":["Postal Item"],"description":"Получение всех почтовых отделений","operationId":"getPostOfficeAll","responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"}}}},"/postalitem/{id}":{"get":{"tags":["Postal Item"],"description":"Получение почтового отправления по id","operationId":"getPostalItemById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostalDelivery"}}}},"401":{"description":"Unauthorized"}}}},"/postalitem/all":{"get":{"tags":["Postal Item"],"description":"Получение списка всех почтовых отправлений","operationId":"getPostalItemAll","responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"}}}},"/postoffice/delete/{id}":{"delete":{"tags":["Postal Item"],"description":"Удаление почтового отделения по id","operationId":"deletePostOfficeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation"},"401":{"description":"Unauthorized"}}}},"/postalitem/delete/{id}":{"delete":{"tags":["Postal Item"],"description":"Удаление почтового отправления по id","operationId":"deletePostalItemById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad request"},"200":{"description":"Successful operation"},"401":{"description":"Unauthorized"}}}}},"components":{"schemas":{"StatusDto":{"type":"object","properties":{"postalId":{"type":"integer","format":"int64"},"postId":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["REGISTRATION","ARRIVAL","DEPARTURE","RECEIVING"]}},"description":"Dto для изменения статуса почтового отправления"},"PostOffice":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"postalCode":{"type":"string"},"address":{"type":"string"}}},"PostalDelivery":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["LETTER","PACKAGE"]},"postalCode":{"type":"string"},"address":{"type":"string"},"recipientName":{"type":"string"},"currentStatus":{"type":"string","enum":["REGISTRATION","ARRIVAL","DEPARTURE","RECEIVING"]},"dateTime":{"type":"string","format":"date-time"},"postOffice":{"$ref":"#/components/schemas/PostOffice"}}},"PostOfficeDto":{"type":"object","properties":{"name":{"type":"string"},"postalCode":{"type":"string"},"address":{"type":"string"}},"description":"Dto для создания нового почтового отделения"},"PostalDeliveryDto":{"type":"object","properties":{"type":{"type":"string","enum":["LETTER","PACKAGE"]},"postalCode":{"type":"string"},"address":{"type":"string"},"recipientName":{"type":"string"},"postId":{"type":"integer","format":"int64"}},"description":"Dto для создания нового почтового отправления"}}}}